requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
idna==3.3
paho-mqtt==1.6.1
psycopg2-binary==2.9.2
pycryptodome==3.11.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.3
urllib3==1.26.7
websocket-client==1.2.1
